pyproject.toml
[project]
name = "home-assistant-remote-commands"
version = "0.1.0"
description = ""
authors = [
    {name = "Michał Wadas",email = "michalwadas@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "ha-mqtt-discoverable (>=0.18.0,<0.19.0)",
    "typer[all] (>=0.15.2,<0.16.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"

[tool.mypy]
strict = true
warn_unused_configs = true
warn_unused_ignores = true
warn_unreachable = true
show_error_codes = true
check_untyped_defs = true
disallow_any_unimported = true
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = false
disallow_any_generics = true
no_implicit_optional = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
implicit_reexport = false
strict_equality = true

[[tool.mypy.overrides]]
module = ["ha_mqtt_discoverable.*"]
follow_untyped_imports = true
